backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.4.2
python-dotenv==1.0.0
bcrypt==4.0.1
jwt==1.3.1
python-jose[cryptography]==3.3.0
cryptography==41.0.7
python-multipart==0.0.6 